# contracts/openapi/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Company Technology Intelligence Platform API
  version: 0.2.0
  description: Phase 2 integration contract for fact APIs and basic RBAC.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /projects:
    get:
      responses:
        '200':
          $ref: '#/components/responses/EnvelopeProjectPage'
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '200':
          $ref: '#/components/responses/EnvelopeProjectDetail'
  /projects/{id}:
    get:
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/EnvelopeProjectDetail'
    patch:
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/EnvelopeProjectDetail'
...
  /documents/{id}:
    get:
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/EnvelopeDocumentDetail'
  /documents/{id}/versions:
    get:
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          $ref: '#/components/responses/EnvelopeDocumentVersions'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    ProjectCreate:
      type: object
      required: [organization_id, code, name]
      properties:
        organization_id: { type: string, format: uuid }
        department_id: { type: [string, 'null'], format: uuid }
        owner_id: { type: [string, 'null'], format: uuid }
        code: { type: string }
        name: { type: string }
        description: { type: [string, 'null'] }
        status: { type: string }
        classification: { type: [string, 'null'] }
    Envelope:
      type: object
      required: [code, data, message, trace_id]
      properties:
        code: { type: integer }
        data: {}
        message: { type: string }
        trace_id: { type: string }
  responses:
...
    EnvelopeProjectPage:
      description: Project page
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    EnvelopeProjectDetail:
      description: Project detail
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
...
    EnvelopeDocumentDetail:
      description: Document detail
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    EnvelopeDocumentVersions:
      description: Document versions
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
...